requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
python-dotenv
pydantic>=2.0
openai
google-genai
httpx
tenacity
